requirements.txt
numpy==1.19.1
SimpleITK==2.0.1
scikit-image==0.17.2
opencv-python==4.2.0.34
matplotlib==3.3.0
Pillow==7.0.0
jupyter